test: Restore original HW Back button sample
